Clean up of the function framework and
corresponding unit tests w.r.t. clang-tidy suggestions.
Furthermore documentation is improved and unnecessary documentation is
removed elsewhere.

part of #682

+++ src/drt_particle_algorithm/particle_dirichlet_bc.cpp
@@ -135,7 +135,7 @@ void PARTICLEALGORITHM::DirichletBoundaryConditionHandler::EvaluateDirichletBoun
     int statedim = container->GetStateDim(PARTICLEENGINE::Position);
 
     // safety check
-    if (statedim != function.NumberComponents())
+    if (static_cast<std::size_t>(statedim) != function.NumberComponents())
       dserror("dimension of function defining dirichlet boundary condition not correct!");
 
     // iterate over owned particles of current type
